backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Maintainer <maintainer@thisproject.org>"]
readme = "README.md"
packages = [{include="app", from="."}]

[tool.poetry.scripts]
webapp-dev = "app.adapter.api:start_web_app"
socketio-app-dev = "app.adapter.sio_srv:entry"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.103.2"
hypercorn = "^0.14"
python-socketio = "^5.10.0"
aiohttp = "^3.9.1"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25"

[tool.poetry.group.dev]

[tool.poetry.group.dev.dependencies]
ruff = "0.1.0"
black = "23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
